+++ src/main/java/ru/innopolis/stc16/innobazaar/entity/User.java
@@ -157,7 +157,7 @@ public class User implements Serializable {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         User user = (User) o;
-        return id == user.id;
+        return id.longValue() == user.id.longValue();
     }
 
     @Override
